video-upsample/requirements.txt
absl-py==1.0.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
argcomplete==2.0.0
asgiref==3.3.4
astunparse==1.6.3
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
av==10.0.0
backcall==0.2.0
base58==2.1.1
bitarray==1.2.2
black==22.3.0
brotlipy @ file:///D:/bld/brotlipy_1648854327487/work
cachetools==4.2.4
certifi==2019.3.9
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==1.1.1
colorama==0.4.4
cryptography==3.4.7
cycler==0.10.0
cytoolz==0.11.2
decorator==5.1.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.14
django-cors-headers==3.7.0
django-debug-toolbar==1.10.1
django-filter==2.4.0
django-heroku==0.3.1
django-rest-knox==4.1.0
django-webpack-loader==1.1.0
djangorestframework==3.12.4
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
facexlib==0.2.5
filterpy==1.4.5
flatbuffers==2.0
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
frozenlist==1.3.0
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
google-search-results==2.4.0
graphics.py==5.0.1.post1
grpcio==1.43.0
gunicorn==20.1.0
h5py==3.6.0
hexbytes==0.2.2
idna==2.8
importlib-metadata==4.10.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
ipython==7.30.1
jedi==0.18.1
joblib==1.1.0
jsonschema==4.5.1
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libclang==12.0.0
llvmlite==0.39.1
lmdb==1.4.0
lru-dict==1.1.7
Markdown==3.3.6
matplotlib==3.4.2
matplotlib-inline==0.1.3
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
mmcv==1.7.0
mmcv-full==1.7.1
mmedit==0.16.0
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
netaddr==0.8.0
numba==0.56.4
numpy==1.20.3
oauthlib==3.0.1
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
parsimonious==0.8.1
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.4.0
pipx==1.0.0
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.24
protobuf==3.19.1
psycopg2==2.9.1
psycopg2-binary==2.8.6
py==1.11.0
py-solc-x==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pycryptodome==3.14.1
Pygments==2.10.0
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1663846997386/work
pyparsing==2.4.7
pyrsistent==0.18.1
PySocks @ file:///D:/bld/pysocks_1661604991356/work
pytest==7.1.2
python-dateutil==2.8.1
python-decouple==3.1
python-dotenv==0.20.0
python3-openid==3.1.0
pytz==2018.5
pywin32==304
PyYAML==6.0
regex==2022.10.31
requests==2.21.0
requests-oauthlib==1.2.0
rlp==2.0.1
rsa==4.8
scikit-learn==1.0.1
scipy==1.6.3
semantic-version==2.9.0
six==1.16.0
sklearn==0.0
sqlparse==0.2.4
stripe==2.27.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
tensorflow-probability==0.8.0rc0
termcolor==1.1.0
threadpoolctl==3.0.0
tomli==2.0.1
toolz==0.11.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.24.2
userpath==1.8.0
varint==1.0.2
wcwidth==0.2.5
web3==5.29.0
websockets==9.1
Werkzeug==2.0.2
whitenoise==3.3.1
win-inet-pton @ file:///D:/bld/win_inet_pton_1667051142467/work
wincertstore==0.2
wrapt==1.13.3
yapf==0.32.0
yarl==1.7.2
zipp==3.6.0
